Eric/update script php

+++ files/script-php.sh
@@ -5,29 +5,36 @@ source "$FUNCTIONS_FILE"
 version_new=8.3
 version_old=8.1
 
-dial " * Installation de  la version ${version_new}." "-"
+dial " * Installation de php avec la version ${version_new}." "-"
 
-## Save existing php package list to packages.txt file
-sudo dpkg -l | grep php | tee packages.txt
+php_version=$(php --version | head -n 1 | cut -d " " -f 2)
 
-# Add Ondrej's PPA
-sudo add-apt-repository ppa:ondrej/php # Press enter when prompted.
-sudo apt update
+if [[ $php_version == "$version_new"* ]]; then
+    dial "La version de PHP est $php_version. Elle est à jour."
+else
 
-# Install new PHP ${version_new} packages
-sudo apt install php${version_new} php${version_new}-cli php${version_new}-{bz2,curl,mbstring,intl}
+    ## Save existing php package list to packages.txt file
+    sudo dpkg -l | grep php | tee packages.txt
 
-# Install FPM OR Apache module
-sudo apt install php${version_new}-fpm
-# OR
-# sudo apt install libapache2-mod-php${version_old}
+    # Add Ondrej's PPA
+    sudo add-apt-repository ppa:ondrej/php # Press enter when prompted.
+    sudo apt update -y
 
-# On Apache: Enable PHP ${version_new} FPM
-sudo a2enconf php${version_new}-fpm
-# When upgrading from an older PHP version:
-sudo a2disconf php${version_old}-fpm
+    # Install new PHP ${version_new} packages
+    sudo apt install php${version_new} php${version_new}-cli php${version_new}-{bz2,curl,mbstring,intl} -y
 
-## Remove old packages
-sudo apt purge php${version_old}*
+    # Install FPM OR Apache module
+    sudo apt install php${version_new}-fpm -y
+    # OR
+    # sudo apt install libapache2-mod-php${version_old}
 
+    # On Apache: Enable PHP ${version_new} FPM
+    sudo a2enconf php${version_new}-fpm -y
+    # When upgrading from an older PHP version:
+    sudo a2disconf php${version_old}-fpm -y
+
+    ## Remove old packages
+    sudo apt purge php${version_old}*
+
+fi
 pause s 5
